Revert "Merge branch 'master' of http://github.com/meloniwastaken/GitDAOTaxi"
This reverts commit da53c050053fe1d2ab9d1b01cc6afe78a56458dc, reversing
changes made to e9b4401cfdb8ec35be022bf9e0a89917439de1fd.

+++ src/ats/controllo/servlet/DeleteServlet.java
@@ -25,7 +25,6 @@ public class DeleteServlet extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)	throws ServletException, IOException {
 		Long id = Long.parseLong(request.getParameter("id"));
-
 		IDAOUtente daoUtente = new DAOUtente();
 		try {
 			Utente u = daoUtente.findById(id);
@@ -40,7 +39,7 @@ public class DeleteServlet extends HttpServlet {
 			System.out.println(e.getMessage());
 		}
 		
-		if (id.longValue() == ((Long) request.getSession().getAttribute("id")).longValue()) {
+		if (id == (Long) request.getSession().getAttribute("id")) {
 			request.getSession().invalidate();
 			response.sendRedirect("/GitDAOTaxi/index.jsp");
 		} else {
